requirements.txt
django
django-cors-headers
djangorestframework
stripe
boto3==1.26.59
django-storages==1.13.2
gunicorn==20.1.0
pillow
